# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/auth": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "获取 token",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/captcha": {
            "get": {
                "description": "通过 github.com/dchest/captcha 插件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "验证码"
                ],
                "summary": "获取验证码 id",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/captcha/:captchaId": {
            "get": {
                "description": "通过 github.com/dchest/captcha 插件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "验证码"
                ],
                "summary": "获取验证码图片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "验证码id",
                        "name": "captchaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    }
}